feat: add login block reflection on login state

+++ youtube-client/src/app/core/components/header/auth/auth.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { LoginService } from 'src/app/auth/services/login.service';
 import { Router } from '@angular/router';
+import { AuthService } from 'src/app/auth/services/auth.service';
 
 
 @Component({
@@ -9,13 +9,15 @@ import { Router } from '@angular/router';
   styleUrls: ['./auth.component.scss']
 })
 export class AuthComponent {
+  public logName: string;
 
-  public name = 'Your Name';
-
-  constructor(public loginService: LoginService, public router: Router) {}
+  constructor( public router: Router, public authService: AuthService ) {}
 
   logout() {
-    localStorage.clear();
-    this.router.navigateByUrl('/auth');
+    if (this.authService.logged) {
+      localStorage.clear();
+      this.authService.logOut();
+      this.router.navigateByUrl('/auth');
+    }
   }
 }
